feat: activate validator and fix some bug
